Moved basic order pages to separate controller
