+++ src/main/java/com/tp32/ecommerceplatform/controller/IndexController.java
@@ -1,27 +1,15 @@
 package com.tp32.ecommerceplatform.controller;
 
-import java.util.List;
-
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestParam;
-
 import com.tp32.ecommerceplatform.dto.LoginDto;
 import com.tp32.ecommerceplatform.dto.RegisterDto;
-import com.tp32.ecommerceplatform.model.Order;
-import com.tp32.ecommerceplatform.model.User;
-import com.tp32.ecommerceplatform.service.OrderService;
 
 @Controller
 public class IndexController {
 
-    private OrderService orderService;
-
-    public IndexController(OrderService orderService) {
-        this.orderService = orderService;
+    public IndexController() {
     }
 
     /*
@@ -67,29 +55,5 @@ public class IndexController {
     @GetMapping("/contact")
     public String contact() {
         return "contact.html";
-    }
-
-    @GetMapping("/orders")
-    public String orders(Model model) {
-        model.addAttribute("status", orderService.getStatus());
-        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        if (!(principal instanceof User))
-            return null;
-
-        model.addAttribute("reverseSortDir", "asc");
-        model.addAttribute("orders", orderService.getOrders(((User) principal).getID()));
-        return "orders.html";
-    }
-
-    @GetMapping("/orders/{sort}")
-    public String orders(Model model, @PathVariable("sort") String sort, @RequestParam("sortDir") String sortDir) {
-        User user = ((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal());
-        if (sort.equals("name"))
-            sort = "user.firstName";
-        List<Order> orders = orderService.getOrdersWithSort(user.getID(), sort, sortDir);
-        model.addAttribute("sortDir", sortDir);
-        model.addAttribute("reverseSortDir", sortDir.equals("asc") ? "desc" : "asc");
-        model.addAttribute("orders", orders);
-        return "orders.html";
     }
 }
